requirements.txt
google-api-python-client==2.149.0
youtube-transcript-api==0.6.3
anthropic==0.42.0
python-dotenv==1.0.1
